+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Features
 
 * Tune up comments page on admin site.
 * Provided ``storage`` directory to store logs, coverage reports and so on.
+* Provided custom command to make messages.
 
 
 Trivial/Internal Changes